Awibof 7492 gc rba locking logic improvement (#607)
* [AWIBOF-7492] Set GC threshold dynamically


* [AWIBOF-7492] Reduce competition by checking RBA validity in the process of acquiring ownership for GC map update


* [AWIBOF-7492] Modified to allocate multiple buffers at once during GC write operation


* [AWIBOF-7492] GC ownership acquisition logic improvement


* [AWIBOF-7492] modified to return buffer list to bufferpool at once


* [AWIBOF-7492] Adjust buffer size for flow control between GC read and write


* [AWIBOF-7492] Adjust buffer size for flow control between GC read and write2


* [AWIBOF-7492] set gc threshold from config


* [AWIBOF-7492] set gc threshold from config (2)

* [AWIBOF-7492] apply review comments

* [AWIBOF-7492] Slow down GC read operation when GC write is busy


* [AWIBOF-7492] fixed memory issue during rba owner recording

* [AWIBOF-7492] remove rba owner recording logic

* [AWIBOF-7492] apply review comments

* [AWIBOF-7492] logging added for rba owner


* [AWIBOF-7492] apply review comments

+++ src/metadata/metadata.cpp
@@ -145,7 +145,7 @@ Metadata::~Metadata(void)
 int
 Metadata::Init(void)
 {
-    int eventId = static_cast<int>(EID(MOUNT_ARRAY_DEBUG_MSG));
+    int eventId = EID(MOUNT_ARRAY_DEBUG_MSG);
     int result = 0;
 
     std::string arrayName = arrayInfo->GetName();
@@ -201,7 +201,7 @@ Metadata::Init(void)
 void
 Metadata::Dispose(void)
 {
-    int eventId = static_cast<int>(EID(UNMOUNT_ARRAY_DEBUG_MSG));
+    int eventId = EID(UNMOUNT_ARRAY_DEBUG_MSG);
     std::string arrayName = arrayInfo->GetName();
 
     POS_TRACE_INFO(eventId, "Start disposing allocator of array {}", arrayName);
@@ -219,7 +219,7 @@ Metadata::Dispose(void)
 void
 Metadata::Shutdown(void)
 {
-    int eventId = static_cast<int>(EID(UNMOUNT_ARRAY_DEBUG_MSG));
+    int eventId = EID(UNMOUNT_ARRAY_DEBUG_MSG);
     std::string arrayName = arrayInfo->GetName();
 
     POS_TRACE_INFO(eventId, "Start shutdown allocator of array {}", arrayName);
@@ -250,7 +250,7 @@ Metadata::NeedRebuildAgain(void)
     }
     else
     {
-        int eventId = static_cast<int>(EID(UNKNOWN_ALLOCATOR_ERROR));
+        int eventId = EID(UNKNOWN_ALLOCATOR_ERROR);
         POS_TRACE_ERROR(eventId, "Can't find context manager to check if rebuild is needed");
         return false;
     }
@@ -272,7 +272,7 @@ Metadata::StopRebuilding(void)
     }
     else
     {
-        int eventId = static_cast<int>(EID(UNKNOWN_ALLOCATOR_ERROR));
+        int eventId = EID(UNKNOWN_ALLOCATOR_ERROR);
         POS_TRACE_ERROR(eventId, "Can't find context manager to check if rebuild is needed");
     }
 }
